# docs/architecture/contracts/rest/openapi.yaml
openapi: 3.1.0
info:
  title: Mintkey Admin REST API
  version: "0.1.0-experimental"
  summary: Operator-facing control-plane HTTP surface for Mintkey.
  description: |
    Mintkey is a credential broker for AI agents. This API is consumed by:
      - The Admin Console (AdminJS) for sensitive operations.
      - Operator-owned tooling (CI, scripts) using API tokens.
      - Internal subscribers calling `/v1/changes` for reconciliation
        (per ADR-0010 change-channel model).

    Stability tier: **experimental**. Breaking changes will bump the major
    `info.version`; surface stability is per-resource and called out in each
    operation `description` where it differs from the document-level tier.

    ## ID conventions

    All identifiers are **ULIDs** (Crockford base32, 26 characters) prefixed
    by their resource type. Tenants are addressed at the wire by their prefixed ULID `tenant_…` on
    every entity for RLS purposes (ADR-0008).

    | Prefix          | Resource                          |
    | --------------- | --------------------------------- |
    | `tenant_…`      | Tenant                            |
    | `operator_…`    | Operator                          |
    | `agent_…`       | Agent                             |
    | `svc_…`         | Service                           |
    | `cred_…`        | Credential (key-version envelope) |
    | `perm_…`        | PermissionGrant                   |
    | `audit_…`       | AuditEvent                        |
...

...
servers:
...
security:
  - OperatorSession: []
  - OperatorBearer: []
paths:
...
  /v1/tenants/{tenant_id}/agents/{agent_id}/permissions:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/AgentId"
    get:
      tags: [Permissions]
      operationId: listPermissions
      summary: List permission grants for an agent.
      description: |
        Returns all grants for the agent scoped to the caller's tenant (RLS).
        No free-text q param (permission_grants has no text fields suitable
        for search). Use service_id to filter by service.
      parameters:
        - in: query
          name: service_id
          required: false
          schema:
            type: string
          description: >
            Filter to grants for a specific service. Accepts UUID or
            svc_<ULID> wire-ID (ADR-0017).
      responses:
        "200":
          description: List of permission grants.
          content:
            application/json:
              schema:
                type: object
                required: [grants]
                properties:
                  grants:
                    type: array
                    items:
                      $ref: "#/components/schemas/PermissionGrant"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Permissions]
      operationId: grantPermission
      summary: Grant a permission to the agent.
      description: |
        The grant is a tuple `(agent, service, action, optional constraints)`.
        The service must belong to the same tenant. Idempotent: granting an
        existing tuple returns 200 with the existing record (not 409).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantPermissionRequest"
            example:
              service_id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S1"
              action: "read:contacts"
              constraints:
                rate_limit:
                  per_minute: 60
      responses:
        "201":
          description: Permission granted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionGrant"
              example:
                id: "perm_01HX5J9F8V8H8V0CG3F2Y5J6P1"
                tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                agent_id: "agent_01HX5J9F8V8H8V0CG3F2Y5J6A1"
                service_id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S1"
                action: "read:contacts"
                constraints:
                  rate_limit:
                    per_minute: 60
                created_at: "2026-05-10T14:00:00Z"
                updated_at: "2026-05-10T14:00:00Z"
        "200":
          description: Permission already exists (idempotent).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionGrant"
              example:
                id: "perm_01HX5J9F8V8H8V0CG3F2Y5J6P1"
                tenant_id: "tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
                agent_id: "agent_01HX5J9F8V8H8V0CG3F2Y5J6A1"
                service_id: "svc_01HX5J9F8V8H8V0CG3F2Y5J6S1"
                action: "read:contacts"
                constraints:
                  rate_limit:
                    per_minute: 60
                created_at: "2026-05-09T10:00:00Z"
                updated_at: "2026-05-09T10:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/tenants/{tenant_id}/agents/{agent_id}/permissions/{permission_id}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/AgentId"
      - in: path
        name: permission_id
        required: true
        schema:
          type: string
          pattern: "^perm_[0-9A-HJKMNP-TV-Z]{26}$"
        description: Permission grant identifier.
    delete:
      tags: [Permissions]
      operationId: revokePermission
      summary: Revoke a permission.
      description: |
        Removes the grant. Outstanding tokens already issued for this
        `(agent, service, action)` tuple remain valid until `exp`; revoke
        the token via the agent revocation endpoint if you need immediate
        effect.
      responses:
        "204":
          description: Revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  ###########################################################################
  # Service API Keys  (ADR-0018; long-lived-api-keys)
  ###########################################################################
...
components:
...
  securitySchemes:
    OperatorSession:
      type: apiKey
      in: cookie
      name: mintkey_session
      description: |
        HttpOnly Secure SameSite=Strict session cookie issued by
        `/v1/auth/internal-login` or `/v1/auth/callback`.
    OperatorBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Operator-owned API token (PAT) for machine-to-machine calls. Issued
        from the Admin Console; not exposed in this OpenAPI document. Sent
        as `Authorization: Bearer …`.
    AgentApiKey:
      type: http
      scheme: bearer
      bearerFormat: mintkey-agent-api-key
      description: |
        Long-lived Agent API Key (`mk_live_…`). Used by the few agent-facing
        REST endpoints; most agent traffic flows through MCP instead.

    AdminUiSignedRequest:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Ed25519 JWT signed by the AdminJS service. Required on every endpoint
        AdminJS calls on behalf of an operator. Carries claims:
          iss: "mintkey/admin-ui"
          sub: <operator_id>
          tnt: <tenant_id>
          aud: "mintkey/admin-api"
          iat, exp (60s TTL)
          jti (UUID; replay-checked against admin_request_jti per ADR-0016.1)
        FastAPI verifies the signature against AdminJS's public key fetched from
        the Vault Adapter under credential type 'admin_ui_signing_key'.
        Per ADR-0014.6 + ADR-0017.1.
    ServiceIdentity:
...
  parameters:
...
    TenantId:
      in: path
      name: tenant_id
      required: true
      schema:
        type: string
        pattern: "^tenant_[0-9A-HJKMNP-TV-Z]{26}$"
      description: ULID-prefixed tenant identifier.
...
    AgentId:
      in: path
      name: agent_id
      required: true
      schema:
        type: string
        pattern: "^agent_[0-9A-HJKMNP-TV-Z]{26}$"
      description: ULID-prefixed agent identifier.
...
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Bad Request"
            status: 400
            detail: "`base_url` must be an absolute https URL."
            instance: "/v1/tenants/tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9/services"
            "mintkey:code": "validation_failed"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
            errors:
              - field: "base_url"
                message: "must be an absolute https URL"
    Unauthorized:
      description: No or invalid credentials.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Unauthorized"
            status: 401
            detail: "Session cookie missing or invalid."
            instance: "/v1/auth/whoami"
            "mintkey:code": "unauthenticated"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
    Forbidden:
      description: Authenticated but not allowed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Forbidden"
            status: 403
            detail: "Operator role 'Auditor' cannot create services."
            instance: "/v1/tenants/tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9/services"
            "mintkey:code": "permission_denied"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
    NotFound:
      description: Resource does not exist or is not visible to the caller.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: "Not Found"
            status: 404
            detail: "No tenant with id 'tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9'."
            instance: "/v1/tenants/tenant_01HX5J9F8V8H8V0CG3F2Y5J6M9"
            "mintkey:code": "tenant_not_found"
            "mintkey:trace_id": "0af7651916cd43dd8448eb211c80319c"
...
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 7807 problem document with two Mintkey-specific extensions:
        `mintkey:code` (machine-readable error code) and
        `mintkey:trace_id` (W3C `trace-id` for correlation).
      properties:
        type:
          type: string
          format: uri-reference
          default: "about:blank"
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        "mintkey:code":
          type: string
          description: Machine-readable error code (snake_case).
          examples:
            - "tenant_not_found"
            - "permission_denied"
            - "credential_in_use"
            - "duplicate_resource"
            - "validation_failed"
            - "unauthenticated"
            - "rate_limited"
            - "change_channel_unavailable"
            - "vault_unavailable"
            - "oidc_unavailable"
        "mintkey:trace_id":
          type: string
          pattern: "^[0-9a-f]{32}$"
          description: W3C `trace-id` (32 hex chars).
        errors:
          type: array
          description: Optional list of per-field validation errors.
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
            required: [field, message]
      required: [title, status]

    # ----- Common identifiers -----
...
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 UTC timestamp.

    # ----- Tenant -----
...
    PermissionGrant:
      type: object
      properties:
        id:
          type: string
          pattern: "^perm_[0-9A-HJKMNP-TV-Z]{26}$"
        tenant_id:
          type: string
          pattern: "^tenant_[0-9A-HJKMNP-TV-Z]{26}$"
        agent_id:
          type: string
          pattern: "^agent_[0-9A-HJKMNP-TV-Z]{26}$"
        service_id:
          type: string
          pattern: "^svc_[0-9A-HJKMNP-TV-Z]{26}$"
        action:
          type: string
        constraints:
          $ref: "#/components/schemas/Constraints"
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
      required:
        - id
        - tenant_id
        - agent_id
        - service_id
        - action
        - created_at
        - updated_at
    GrantPermissionRequest:
      type: object
      properties:
        service_id:
          type: string
          pattern: "^svc_[0-9A-HJKMNP-TV-Z]{26}$"
        action:
          type: string
        constraints:
          $ref: "#/components/schemas/Constraints"
      required: [service_id, action]

    # ----- Audit -----
...
    Constraints:
      type: object
      additionalProperties: false
      description: |
        ABAC constraints on a permission grant. Closed schema per ADR-0016.4.
      properties:
        rate_limit:
          type: object
          additionalProperties: false
          description: Per-agent rate limit on this (service, action) tuple.
          properties:
            requests_per_second: { type: integer, minimum: 1 }
            burst:               { type: integer, minimum: 1 }
        time_window:
          type: object
          additionalProperties: false
          description: Time-of-day window the permission is valid in.
          properties:
            timezone:    { type: string, description: "IANA tz name, e.g., Europe/Bucharest." }
            days:
              type: array
              items: { type: string, enum: [Mon, Tue, Wed, Thu, Fri, Sat, Sun] }
            start_local: { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$" }
            end_local:   { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$" }
        request_path_prefix:
          type: array
          description: Allowed path prefixes within the registered service.
          items: { type: string }
        source_ip_allowlist:
          type: array
          items: { type: string, format: cidr }
...